Implemented a class that handles the onAuthStateChange listener
