+++ app/src/main/java/com/example/musicmap/util/ui/MMActivity.java
@@ -1,0 +1,55 @@
+package com.example.musicmap.util.ui;
+
+import android.content.Intent;
+import android.os.Bundle;
+import android.os.PersistableBundle;
+
+import androidx.annotation.NonNull;
+import androidx.annotation.Nullable;
+import androidx.appcompat.app.AppCompatActivity;
+
+import com.example.musicmap.screens.auth.AuthActivity;
+import com.google.firebase.auth.FirebaseAuth;
+import com.google.firebase.auth.FirebaseUser;
+
+public class MMActivity extends AppCompatActivity implements FirebaseAuth.AuthStateListener {
+
+    private FirebaseAuth auth;
+
+    @Override
+    protected void onCreate(@Nullable Bundle savedInstanceState) {
+        super.onCreate(savedInstanceState);
+        auth = FirebaseAuth.getInstance();
+    }
+
+    @Override
+    public void onAuthStateChanged(@NonNull FirebaseAuth firebaseAuth) {
+        auth = firebaseAuth;
+        FirebaseUser firebaseUser = auth.getCurrentUser();
+
+        if (firebaseUser == null) {
+            Intent authIntent = new Intent(this, AuthActivity.class);
+            authIntent.setFlags(Intent.FLAG_ACTIVITY_CLEAR_TOP);
+            startActivity(authIntent);
+            finish();
+        }
+    }
+
+    @Override
+    public void onPause() {
+        super.onPause();
+        auth.removeAuthStateListener(this);
+    }
+
+    @Override
+    protected void onResume() {
+        super.onResume();
+        auth.addAuthStateListener(this);
+    }
+
+    @Override
+    public void onStop() {
+        super.onStop();
+        auth.removeAuthStateListener(this);
+    }
+}
